+++ src/main/webapp/dao/InstructorDao.java
@@ -15,6 +15,6 @@ import pojo.Instructor;
  * if anyone wants to add his own customized quarry 
  * please use XXXHibernateDao
  */
-public interface InstructorDao extends GenericDAO <Instructor, Integer>{
+public interface InstructorDao extends GenericDAO <Instructor, Long>{
     public Instructor findByEmail (String email);
 }
