+++ src/main/webapp/dao/TraineeDao.java
@@ -15,7 +15,7 @@ import java.util.List;
  * @author root this interface no one allowed to add any code to if anyone wants
  * to add his own customized quarry please use XXXHibernateDao
  */
-public interface TraineeDao extends GenericDAO<Trainee, Integer> {
+public interface TraineeDao extends GenericDAO<Trainee, Long> {
     public Trainee findByEmail (String email);
 
     public List<Trainee> findByAllActivated ();
